Add files via upload
This is Task 1 of my internship at Txon in Web Development and Design. In this task, I have created a Landing Page of a Cashew's Factory using HTML and CSS. A landing page is one of the best web development projects for beginners. This project demands a foundational understanding of HTML and CSS. Through this task, I understood the basics of HTML and CSS i.e how to add columns, add footers, arrange text and images. Most importantly, I was able to use my creativity to make the page look impressive. The alignments, the padding, the color palette, boxes and all the other elements on the page require attention.